Refactor out functionality into overlaysAccessor.ts
